prototype first map
added a sphere representing a basic enemy, added waypoints that can be used for enemy movement. also added start and end points.

One should be able to implement an algorithm that allows the enemy (sphere) to move towards the end point by following the endpoints. Once this end point is reached, an event must be triggered.